Move template refreshing code into templates app
