+++ temba/templates/models.py
@@ -28,6 +28,42 @@ class Template(models.Model):
     # when this template was created
     created_on = models.DateTimeField(default=timezone.now)
 
+    @classmethod
+    def update_local(cls, channel, wa_templates):
+        from .whatsapp import STATUS_MAPPING, extract_components, parse_language
+
+        channel_namespace = channel.config.get("fb_namespace", "")
+
+        # run through all our templates making sure they are present in our DB
+        seen = []
+        for template in wa_templates:
+            template_status = template["status"].upper()
+            if template_status not in STATUS_MAPPING:  # ignore if this is a status we don't know about
+                continue
+
+            components, all_supported = extract_components(template["components"])
+
+            status = STATUS_MAPPING[template_status]
+            if not all_supported:
+                status = TemplateTranslation.STATUS_UNSUPPORTED_COMPONENTS
+
+            missing_external_id = f"{template['language']}/{template['name']}"
+            translation = TemplateTranslation.get_or_create(
+                channel,
+                template["name"],
+                locale=parse_language(template["language"]),
+                status=status,
+                external_locale=template["language"],
+                external_id=template.get("id", missing_external_id[:64]),
+                namespace=template.get("namespace", channel_namespace),
+                components=components,
+            )
+
+            seen.append(translation)
+
+        # trim any translations we didn't see
+        TemplateTranslation.trim(channel, seen)
+
     def is_approved(self):
         """
         Returns whether this template has at least one translation and all are approved
